+++ explainaboard/builders/qa_multiple_choice.py
@@ -75,7 +75,7 @@ class QAMultipleChoiceExplainaboardBuilder(ExplainaboardBuilder):
         self,
         info: SysOutputInfo,
         system_output_object: Iterable[dict],
-        feature_table: Optional[Table] = {},
+        feature_table: Optional[Table] = None,
         user_defined_feature_config=None,
     ):
         super().__init__(
@@ -172,6 +172,7 @@ class QAMultipleChoiceExplainaboardBuilder(ExplainaboardBuilder):
             self._data[_id] = dict_sysout
             yield _id, dict_sysout
 
+    # TODO(gneubig): should this be generalized or is it task specific?
     def get_overall_performance(self):
         predicted_labels, true_labels = [], []
 
@@ -202,42 +203,6 @@ class QAMultipleChoiceExplainaboardBuilder(ExplainaboardBuilder):
             else:
                 self._info.results.overall[metric_name] = overall_performance
 
-    def _bucketing_samples(self, sysout_iterator):
-
-        sample_address = ""
-        feature_to_sample_address_to_value = {}
-
-        # Preparation for bucketing
-        for _id, dict_sysout in sysout_iterator:
-
-            sample_address = str(_id)  # this could be abstracted later
-            for feature_name in self._info.features.get_bucket_features():
-                if feature_name not in feature_to_sample_address_to_value.keys():
-                    feature_to_sample_address_to_value[feature_name] = {}
-                else:
-                    feature_to_sample_address_to_value[feature_name][
-                        sample_address
-                    ] = dict_sysout[feature_name]
-
-        # Bucketing
-        for feature_name in tqdm(
-            self._info.features.get_bucket_features(), desc="bucketing"
-        ):
-
-            self._samples_over_bucket[feature_name] = eval(
-                self._info.features[feature_name].bucket_info._method
-            )(
-                dict_obj=feature_to_sample_address_to_value[feature_name],
-                bucket_number=self._info.features[feature_name].bucket_info._number,
-                bucket_setting=self._info.features[feature_name].bucket_info._setting,
-            )
-
-            # print(f"self._samples_over_bucket.keys():\n{self._samples_over_bucket.keys()}")
-
-            # evaluating bucket: get bucket performance
-            self._performances_over_bucket[feature_name] = self.get_bucket_performance(
-                feature_name
-            )
 
     def get_bucket_performance(self, feature_name: str):
         """
@@ -304,29 +269,3 @@ class QAMultipleChoiceExplainaboardBuilder(ExplainaboardBuilder):
                 bucket_name_to_performance[bucket_interval].append(bucket_performance)
 
         return sort_dict(bucket_name_to_performance)  # noqa
-
-    def _generate_report(self):
-        dict_fine_grained = {}
-        for feature_name, metadata in self._performances_over_bucket.items():
-            dict_fine_grained[feature_name] = []
-            for bucket_name, bucket_performance in metadata.items():
-                bucket_name = analysis.beautify_interval(bucket_name)
-
-                # instantiation
-                dict_fine_grained[feature_name].append(bucket_performance)
-
-        self._info.results.fine_grained = dict_fine_grained
-
-    def _print_bucket_info(self):
-        for feature_name in self._performances_over_bucket.keys():
-            print_dict(  # noqa
-                self._performances_over_bucket[feature_name], feature_name
-            )
-
-    def run(self) -> SysOutputInfo:
-        eb_generator = self._complete_feature()
-        self._bucketing_samples(eb_generator)
-        self.get_overall_performance()
-        self._print_bucket_info()
-        self._generate_report()
-        return self._info
